move java hash functions together
